backend/requirements.txt
# === requirements.txt ===
fastapi
uvicorn
networkx
pyyaml
python-multipart
pydantic
